feat(TDP-4104): preparation steps vanished     * Do not remove the content (action)  of a step that is reused in another step

+++ dataprep-preparation/src/main/java/org/talend/dataprep/preparation/task/PreparationCleaner.java
@@ -21,6 +21,8 @@ import java.util.stream.Stream;
 
 import javax.annotation.Resource;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.scheduling.annotation.EnableScheduling;
@@ -38,6 +40,8 @@ import org.talend.dataprep.security.SecurityProxy;
 @Component
 @EnableScheduling
 public class PreparationCleaner {
+
+    private static final Logger LOGGER = LoggerFactory.getLogger(PreparationCleaner.class);
 
     @Autowired
     private PreparationRepository repository;
@@ -61,11 +65,13 @@ public class PreparationCleaner {
     @Autowired
     private ForAll forAll;
 
+
     /**
      * Get all the step ids that belong to a preparation
      *
      * @return The step ids
      */
+
     private Set<String> getPreparationStepIds() {
         return repository.list(Preparation.class) //
                 .flatMap(p -> p.getSteps().stream().map(Step::getId)) //
@@ -99,9 +105,17 @@ public class PreparationCleaner {
                     repository.remove(stepToRemove);
 
                     // Remove actions linked to step
-                    final PreparationActions preparationActionsToRemove = new PreparationActions();
-                    preparationActionsToRemove.setId(step.getContent());
-                    repository.remove(preparationActionsToRemove);
+                    // if this step re-use an existing actions we don't delete the actions
+                    boolean criterion = repository.exist(PersistentStep.class,
+                            "content" + "='" + step.getContent() + "'");
+                    if (criterion) {
+                        LOGGER.info("Don't removing step content {} it still used by another step.", step.getContent());
+                    } else {
+                        LOGGER.info("Removing step content {}.", step.getContent());
+                        final PreparationActions preparationActionsToRemove = new PreparationActions();
+                        preparationActionsToRemove.setId(step.getContent());
+                        repository.remove(preparationActionsToRemove);
+                    }
 
                     // Remove metadata linked to step
                     final StepRowMetadata stepRowMetadataToRemove = new StepRowMetadata();
